Moving prefixTokenError to the more general and exported PrefixError
