+++ errors_test.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
+	"golang.org/x/oauth2"
 )
 
 var (
@@ -71,4 +72,37 @@ func TestNewTimeErrorErrorNil(t *testing.T) {
 
 	te := NewTimeError(nil)
 	r.Equal("", te.Error())
+}
+
+func TestPrefixErrorRetrieveError(t *testing.T) {
+	var r = require.New(t)
+
+	var err *oauth2.RetrieveError
+	if err := PrefixError("failed to get access token", err); err != nil {
+		r.Equal("ErrInvalidCredentials: failed to get access token: <nil>", err.Error())
+	} else {
+		t.Error("expected a prefixed error")
+	}
+}
+
+func TestPrefixErrorNil(t *testing.T) {
+	var r = require.New(t)
+
+	var err error
+	if err := PrefixError("failed to get access token", err); err != nil {
+		r.Equal("failed to get access token", err.Error())
+	} else {
+		t.Error("expected a prefixed error")
+	}
+}
+
+func TestPrefixErrorOther(t *testing.T) {
+	var r = require.New(t)
+
+	var err = errTimeErrorUnknown
+	if err := PrefixError("failed to get access token", err); err != nil {
+		r.Equal("failed to get access token: testing error", err.Error())
+	} else {
+		t.Error("expected a prefixed error")
+	}
 }
